SPMI: implement rest of the commands
also return EINVAL if we get passed an unimplemented / unexpected opcode (the previous version wasn't doing this, resulting in "funny" FSM lockups when e.g. a reg 0 write was issued onto the driver).

many of the validations I'm doing in spmi_cmd_read / spmi_cmd_write are possibly not needed, as they would be guaranteed by the upper spmi.c layer, but I'm not 100% sure and better be safe than sorry